feat(bex): move useCreatePool into beraJS hook + move pool naming function definition into it's own file
