+++ packages/berajs/package.json
@@ -23,6 +23,7 @@
     "@bera/config": "workspace:*",
     "@bera/graphql": "workspace:*",
     "@bera/wagmi": "workspace:*",
+    "@berachain-foundation/berancer-sdk": "^0.29.0",
     "@dynamic-labs/sdk-react-core": "^2.5.2",
     "@metamask/browser-passworder": "^5.0.0",
     "@tanstack/react-query": "^5.26.3",
