+++ packages/berajs/src/hooks/modules/dex/index.ts
@@ -2,3 +2,4 @@ export * from "./usePoolRecentSwaps";
 export * from "./usePoolRecentProvisions";
 export * from "./usePoolHistoricalData";
 export * from "./useUserPools";
+export * from "./useCreatePool";
